requirements.txt
aiohappyeyeballs==2.6.1 ; python_version >= "3.12" and python_version < "3.14"
aiohttp==3.12.15 ; python_version >= "3.12" and python_version < "3.14"
aiosignal==1.4.0 ; python_version >= "3.12" and python_version < "3.14"
attrs==25.3.0 ; python_version >= "3.12" and python_version < "3.14"
certifi==2025.8.3 ; python_version >= "3.12" and python_version < "3.14"
charset-normalizer==3.4.3 ; python_version >= "3.12" and python_version < "3.14"
cloudpickle==3.1.1 ; python_version >= "3.12" and python_version < "3.14"
colorama==0.4.6 ; python_version >= "3.12" and python_version < "3.14" and platform_system == "Windows"
contourpy==1.3.3 ; python_version >= "3.12" and python_version < "3.14"
cycler==0.12.1 ; python_version >= "3.12" and python_version < "3.14"
datasets==4.0.0 ; python_version >= "3.12" and python_version < "3.14"
dill==0.3.8 ; python_version >= "3.12" and python_version < "3.14"
filelock==3.19.1 ; python_version >= "3.12" and python_version < "3.14"
fonttools==4.59.2 ; python_version >= "3.12" and python_version < "3.14"
frozenlist==1.7.0 ; python_version >= "3.12" and python_version < "3.14"
fsspec==2025.3.0 ; python_version >= "3.12" and python_version < "3.14"
fsspec[http]==2025.3.0 ; python_version >= "3.12" and python_version < "3.14"
hf-xet==1.1.9 ; python_version >= "3.12" and python_version < "3.14" and (platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "arm64" or platform_machine == "aarch64")
huggingface-hub==0.34.4 ; python_version >= "3.12" and python_version < "3.14"
idna==3.10 ; python_version >= "3.12" and python_version < "3.14"
jinja2==3.1.6 ; python_version >= "3.12" and python_version < "3.14"
joblib==1.5.2 ; python_version >= "3.12" and python_version < "3.14"
kagglehub==0.3.13 ; python_version >= "3.12" and python_version < "3.14"
kiwisolver==1.4.9 ; python_version >= "3.12" and python_version < "3.14"
llvmlite==0.45.0rc1 ; python_version >= "3.12" and python_version < "3.14"
markupsafe==3.0.2 ; python_version >= "3.12" and python_version < "3.14"
matplotlib==3.10.6 ; python_version >= "3.12" and python_version < "3.14"
mpmath==1.3.0 ; python_version >= "3.12" and python_version < "3.14"
multidict==6.6.4 ; python_version >= "3.12" and python_version < "3.14"
multiprocess==0.70.16 ; python_version >= "3.12" and python_version < "3.14"
narwhals==2.3.0 ; python_version >= "3.12" and python_version < "3.14"
networkx==3.5 ; python_version >= "3.12" and python_version < "3.14"
numba==0.62.0rc1 ; python_version >= "3.12" and python_version < "3.14"
numpy==2.3.2 ; python_version >= "3.12" and python_version < "3.14"
nvidia-cublas-cu12==12.8.4.1 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version >= "3.12" and python_version < "3.14"
nvidia-cuda-cupti-cu12==12.8.90 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version >= "3.12" and python_version < "3.14"
nvidia-cuda-nvrtc-cu12==12.8.93 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version >= "3.12" and python_version < "3.14"
nvidia-cuda-runtime-cu12==12.8.90 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version >= "3.12" and python_version < "3.14"
nvidia-cudnn-cu12==9.10.2.21 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version >= "3.12" and python_version < "3.14"
nvidia-cufft-cu12==11.3.3.83 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version >= "3.12" and python_version < "3.14"
nvidia-cufile-cu12==1.13.1.3 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version >= "3.12" and python_version < "3.14"
nvidia-curand-cu12==10.3.9.90 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version >= "3.12" and python_version < "3.14"
nvidia-cusolver-cu12==11.7.3.90 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version >= "3.12" and python_version < "3.14"
nvidia-cusparse-cu12==12.5.8.93 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version >= "3.12" and python_version < "3.14"
nvidia-cusparselt-cu12==0.7.1 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version >= "3.12" and python_version < "3.14"
nvidia-nccl-cu12==2.27.3 ; platform_system == "Linux" and platform_machine != "aarch64" and python_version >= "3.12" and python_version < "3.14"
nvidia-nvjitlink-cu12==12.8.93 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version >= "3.12" and python_version < "3.14"
nvidia-nvtx-cu12==12.8.90 ; platform_system == "Linux" and platform_machine == "x86_64" and python_version >= "3.12" and python_version < "3.14"
packaging==25.0 ; python_version >= "3.12" and python_version < "3.14"
pandas==2.3.2 ; python_version >= "3.12" and python_version < "3.14"
pillow==11.3.0 ; python_version >= "3.12" and python_version < "3.14"
plotly==6.3.0 ; python_version >= "3.12" and python_version < "3.14"
propcache==0.3.2 ; python_version >= "3.12" and python_version < "3.14"
pyarrow==21.0.0 ; python_version >= "3.12" and python_version < "3.14"
pyparsing==3.2.3 ; python_version >= "3.12" and python_version < "3.14"
python-dateutil==2.9.0.post0 ; python_version >= "3.12" and python_version < "3.14"
pytz==2025.2 ; python_version >= "3.12" and python_version < "3.14"
pyyaml==6.0.2 ; python_version >= "3.12" and python_version < "3.14"
requests==2.32.5 ; python_version >= "3.12" and python_version < "3.14"
scikit-learn==1.7.1 ; python_version >= "3.12" and python_version < "3.14"
scipy==1.16.1 ; python_version >= "3.12" and python_version < "3.14"
setuptools==80.9.0 ; python_version >= "3.12" and python_version < "3.14"
shap==0.48.0 ; python_version >= "3.12" and python_version < "3.14"
six==1.17.0 ; python_version >= "3.12" and python_version < "3.14"
slicer==0.0.8 ; python_version >= "3.12" and python_version < "3.14"
sympy==1.14.0 ; python_version >= "3.12" and python_version < "3.14"
threadpoolctl==3.6.0 ; python_version >= "3.12" and python_version < "3.14"
torch==2.8.0 ; python_version >= "3.12" and python_version < "3.14"
torchvision==0.23.0 ; python_version >= "3.12" and python_version < "3.14"
tqdm==4.67.1 ; python_version >= "3.12" and python_version < "3.14"
triton==3.4.0 ; python_version >= "3.12" and python_version < "3.14" and platform_system == "Linux" and platform_machine == "x86_64"
typing-extensions==4.15.0 ; python_version >= "3.12" and python_version < "3.14"
tzdata==2025.2 ; python_version >= "3.12" and python_version < "3.14"
urllib3==2.5.0 ; python_version >= "3.12" and python_version < "3.14"
xgboost==3.0.4 ; python_version >= "3.12" and python_version < "3.14"
xxhash==3.5.0 ; python_version >= "3.12" and python_version < "3.14"
yarl==1.20.1 ; python_version >= "3.12" and python_version < "3.14"
